[BREAKING] add `matrixcontrol` to applicationId
This change was necessary in order to specify that this application isn't "simulatan" but just an application published by me under that group.
Warning: if you already had the app installed, the old one will stay untouched and not receive any updates anymore. Furthermore, preferences will not be kept, although this isn't a problem since the last update broke them anyway.